refactor (NameTags/Enchantments)
