+++ src/main/kotlin/net/ccbluex/liquidbounce/features/module/modules/render/nametags/NametagRenderer.kt
@@ -89,17 +89,7 @@ class NametagRenderer {
 
         // Draw enchantments directly for the entity (regardless of whether items are shown)
         if (NametagShowOptions.ENCHANTMENTS.isShowing() && nametag.entity is LivingEntity) {
-            val entityPos = nametag.entity.pos
-            val worldX = entityPos.x.toFloat()
-            val worldY = (entityPos.y + nametag.entity.height + 0.5f).toFloat()
-
-            NametagEnchantmentRenderer.drawEntityEnchantments(
-                env,
-                nametag.entity,
-                worldX,
-                worldY,
-                fontBuffers
-            )
+            NametagEnchantmentRenderer.drawEntityEnchantments(env, nametag.entity, fontBuffers)
         }
 
         matrixStack.pop()
