+++ src/main/kotlin/net/ccbluex/liquidbounce/features/module/modules/render/nametags/enchantments/EnchantmentDataStructures.kt
@@ -1,0 +1,34 @@
+/*
+ * This file is part of LiquidBounce (https://github.com/CCBlueX/LiquidBounce)
+ *
+ * Copyright (c) 2015 - 2025 CCBlueX
+ *
+ * LiquidBounce is free software: you can redistribute it and/or modify
+ * it under the terms of the GNU General Public License as published by
+ * the Free Software Foundation, either version 3 of the License, or
+ * (at your option) any later version.
+ *
+ * LiquidBounce is distributed in the hope that it will be useful,
+ * but WITHOUT ANY WARRANTY; without even the implied warranty of
+ * MERCHANTABILITY or FITNESS FOR A PARTICULAR PURPOSE. See the
+ * GNU General Public License for more details.
+ *
+ * You should have received a copy of the GNU General Public License
+ * along with LiquidBounce. If not, see <https://www.gnu.org/licenses/>.
+ */
+package net.ccbluex.liquidbounce.features.module.modules.render.nametags.enchantments
+
+import net.ccbluex.liquidbounce.render.engine.font.processor.TextProcessor.ProcessedText
+
+ // Represents a single enchantment cell in the display grid
+data class EnchantCell(
+    val processedText: ProcessedText,
+    val textWidth: Float,
+    val isCurse: Boolean
+)
+
+ // Represents a column of enchantment cells
+data class EnchantColumn(
+    val cells: List<EnchantCell>,
+    val width: Float
+)
